redisqueue deployment test to heroku
