--- Names2.csv
@@ -1,2 +1,0 @@
-from_id,from_login,from_name,to_id,to_login,to_name,followed_at,closure,follow_total,triad_set,k-connected,total-connected
-129447822,boddwvlqforc,boddwvlqforc,49886567,kawyua,kawyua,2016-07-29T15:53:03Z,0,500,[],[],0
