+++ Pipfile
@@ -24,6 +24,8 @@ MarkupSafe = "==1.1.1"
 SQLAlchemy = "==1.3.7"
 Werkzeug = "==0.15.5"
 python-dotenv = "*"
+rq = "*"
+redis = "*"
 
 [dev-packages]
