+++ templates/add_task.html
@@ -1,0 +1,63 @@
+<div class="container">
+  <div class="row">
+    <div class="col">
+      <h5 class="mt-3 mb-3">Job queue</h5>
+      {% if jobs %}
+      {% for job in jobs %}
+      <div class="card mb-3">
+        <div class="card-body">
+          <h6>{{ job.func_name }}</h6>
+          <small class="text-muted d-block">Job ID: {{ job.id }}</small>
+          <small class="text-muted d-block">Status: {{ job.status }}</small>
+          <small class="text-muted d-block">Created at: {{ job.created_at.strftime('%a, %d %b %Y %H:%M:%S') }}</small>
+          <small class="text-muted d-block">Enqueued at: {{ job.enqueued_at.strftime('%a, %d %b %Y %H:%M:%S') }}</small>
+          <small class="text-muted d-block">Follows Queued: {{ job.description }}</small>
+        </div>
+      </div>
+      {% endfor %}
+      {% else %}
+      <p>No other jobs in the queue. Queueing follow history for {{userdata[0]["login"]}}</p>
+      {% endif %}
+      <div class="card green">
+        <div class="additional">
+          <div class="user-card">
+            <div class="level center">
+              Views: {{ userdata[0]["view_count"] }}
+            </div>
+            <div class="points center">
+              Inserting: {{ task.description }}
+            </div>
+            <a href='https://twitch.tv/{{userdata[0]["login"]}}/profile'><img src={{userdata[0]["profile_image_url"]}} alt="test" class="avatar center"></a>
+            <svg width="110" height="110" viewBox="0 0 250 250" xmlns="http://www.w3.org/2000/svg" role="img" aria-labelledby="title desc" class="center">
+               <circle cx="125" cy="125" r="120" fill="rgba(0,0,0,0.15)" />
+    
+            </svg>
+          </div>
+          <div class="more-info">
+            <h2 class= "text-center"  ><a href="https://twitch.tv/{{ userdata[0]['login'] }}"  target="_blank" rel="noopener noreferrer">{{ userdata[0]["login"] }}</a>
+              <button class="btn btn-success2 btn-circle btn-circle-sm m-1" onclick='adduser("{{ userdata[0]["login"]}}", {{ userdata[0]["id"] }})'>+</button></h2>
+            <div class="coords">
+              <h3>{{userdata[0]["broadcaster_type"]}}</h3>
+              <h3>Joined {{ userdata[0]["created_at"] }}</h3>
+              <small class="text-muted d-block">Enqueued at: {{ task.enqueued_at.strftime('%a, %d %b %Y %H:%M:%S') }}</small>
+            </div>
+            <div class="coords">
+              
+              <h3>{{ userdata[0]['type']}}</h3>
+              <h3>  <a href="https://twitch.tv/{{userdata[0]['login']}}">{{userdata[0]["login"]}}🔗</a></h3>
+            </div>
+            <div class="stats">
+              
+            </div>
+          </div>
+        </div>
+        <div class="general">
+          <h2 class= "text-center"  ><a href="https://twitch.tv/{{ userdata[0]['login'] }}"  target="_blank" rel="noopener noreferrer">{{ userdata[0]["login"] }}</a>
+            <button class="btn btn-success2 btn-circle btn-circle-sm m-1" onclick='adduser("{{ userdata[0]["login"]}}", {{ userdata[0]["id"] }})'>+</button></h2>
+          <p>{{ userdata[0]["description"] }}</p>
+          <span class="more">Mouse over the card for more info</span>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
