+++ templates/follow.html
@@ -95,8 +95,7 @@
                 <a>{{triad[0]}}, </a>
                 {% elif loop.index-1 == 5 %}
                 <a data-toggle="tooltip" data-placement="top" 
-                title="{{triad[0]}},{% else %}{{triad[0]}},{% endif %}{% endfor %}{% if loop.index-1 >= 5 %}">
-                and {{followdata[loop.index-1]["triad"]|length - 5}} others</a>
+                title="{{triad[0]}},{% else %}{{triad[0]}},{% endif %}{% endfor %}{% if loop.index-1 >= 5 %}"> and {{followdata[loop.index-1]["triad"]|length - 5}} others</a>
             {% endif %}
           {% else %}
             {% for triad in followdata[loop.index-1]["triad"] %}
@@ -112,8 +111,7 @@
                 <a>{{commonfollowsession[0]}}, </a>
                 {% elif loop.index-1 == 5 %}
                 <a data-toggle="tooltip" data-placement="top" 
-                title="{{commonfollowsession[0]}},{% else %}{{commonfollowsession[0]}},{% endif %}{% endfor %}{% if loop.index-1 >= 5 %}">
-                and {{followdata[loop.index-1]["commonfollowsession"]|length - 5}} others</a>
+                title="{{commonfollowsession[0]}},{% else %}{{commonfollowsession[0]}},{% endif %}{% endfor %}{% if loop.index-1 >= 5 %}"> and {{followdata[loop.index-1]["commonfollowsession"]|length - 5}} others</a>
             {% endif %}
           {% else %}
             {% for commonfollowsession in followdata[loop.index-1]["commonfollowsession"] %}
