+++ templates/graph.html
@@ -2,32 +2,45 @@
 <meta id="my-data" data-name="graphdata" data-other="{{graphdata}}">
 <div id="cy"></div>
 <script>
-  var graphdata = {{ graphdata|tojson|safe }};
-  var total = Math.floor({{ total }});
-  graphdata = jQuery.parseJSON(graphdata);
-  console.log(graphdata);
+  var followdata = {{ followdata|tojson|safe }};
+  console.log(followdata)
+  var users = {{ data|tojson|safe }};
+  console.log(users)
+  var userdata = {{ userdata|tojson|safe }};
+  console.log(userdata)
+  var total = Math.floor({{ firstfollow }});
   console.log(total);
-  var followdata = graphdata.follows;
-  var users = graphdata.users;
+  console.log(userdata[0]["login"])
   triad = 0;
   var totalparts = 30;
   var partition = total/totalparts;
+  console.log(partition)
   promise = new Promise(function(resolve, reject) {
     data = []
-    for (var i = 0; i < users.length; i++) {
-          user = users[i].id;
-          login = users[i].login;
-          profile_image_url = users[i].profile_image_url;
-          data.push({
-              data: {id: user, name: login, profile_image_url: profile_image_url}
-              }
-          );
-    };
+    user = userdata[0]["id"];
+    login = userdata[0]["login"];
+    profile_image_url = userdata[0]["profile_image_url"];
+    data.push({
+        data: {id: user, name: login, profile_image_url: profile_image_url, time: "0s"}
+        }
+    );
     for (var i = 0; i < followdata.length; i++) {
-      var source = followdata[i].from_id;
-      var target = followdata[i].to_id;
-      var followtime = followdata[i].followtime;
+      var source = followdata[i]["from_id"];
+      var target = followdata[i]["to_id"];
+      var followtime = followdata[i]["followtime"];
       var time = Math.floor(followtime/partition).toString(10)+'s';
+      console.log(time)
+      user = users[i]["id"];
+      login = users[i]["login"];
+      profile_image_url = users[i]["profile_image_url"];
+      data.push({
+          data: {
+            id: user, 
+            name: login, 
+            profile_image_url: profile_image_url,
+            time: time
+          }
+      });
       data.push({
           data: {
               id: 'edge' + i,
@@ -39,10 +52,10 @@
           }
       });
       if (followdata[i].hasOwnProperty("triad")){
-        for(var j = 0; j < followdata[i].triad.length; j++) {
-          var source = followdata[i].triad[j][1];
-          var target = followdata[i].to_id;
-          var followedtime = followdata[i].triad[j][2];
+        for(var j = 0; j < followdata[i]["triad"].length; j++) {
+          var source = followdata[i]["triad"][j][1];
+          var target = followdata[i]["to_id"];
+          var followedtime = Math.max(followdata[i]["triad"][j][2], followtime);
           var time = Math.floor(followedtime/partition).toString(10)+'s';
           data.push({
               data: {
@@ -50,7 +63,7 @@
                   source: source.toString(10),
                   target: target.toString(10),
                   color: '#aa5f85',
-                  width: 5,
+                  width: 8,
                   time: time
               }
           });
@@ -80,7 +93,10 @@
           'content': 'data(name)',
           'font-size': "36px",
           'background-color': '#30c9bc',
-          'background-image': 'data(profile_image_url)'
+          'background-image': 'data(profile_image_url)',
+          'opacity': 0,
+          'transition-duration': '2s',
+          'transition-delay': 'data(time)'
         }
 
       },
@@ -101,16 +117,14 @@
         selector: '.highlighted',
         style: {
           'opacity': 0.8,
-          'width': 17,
-          'line-color': 'data(color)',
-          'transition-property': 'opacity, width'
+          'transition-property': 'opacity'
         }
       }
     ],
 
     elements: data
   });
-  var bfs = cy.edges().addClass('highlighted');
+  var bfs = cy.elements().addClass('highlighted');
 });
 var now = Date.now();
 console.log(now);
